+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Record of the modification in project development.
 ## [Unreleased] - yyyy-mm-dd
 ### Added
 - [Issue 31 : Add extra partition functionality.](https://github.com/suikan4github/kaiten-yaki/issues/31)
-- [Issue 35 : Support "M/G/T" as size prefix.](https://github.com/suikan4github/kaiten-yaki/issues/35)
+- [Issue 33 : Support "M/G/T" as size prefix.](https://github.com/suikan4github/kaiten-yaki/issues/33)
 
 ### Changed
 ### Deprecated